requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pydeck>=0.8.0
plotly>=5.18.0
numpy>=1.24.0
gspread>=5.12.0
google-auth>=2.23.0
